trying to get the delete function working
keeps coming up with deleteItem() undefined even though the function is defined in that for loop. Not sure if you have deleted all the items from teh API somehow, and that is the problem with the error element not defined. because then when you add a new item to the list then this error doesn't come up anymore. So I think because all the elements were deleted it was not being defined (i think)