working delete button and working clear all list
you were nearly there but you just needed to replace all the i['name'],i['desc'] stuff with just the id. So actually I was overcomplicating it. then the clear all button just needed to be adjusted slightly as it was only deleting the items locally, but now it deletes the items from the API. also you learned that i['id'] can alternatively be written as i.id