This gives the error in the console net:: ERR_CONECTION_RESET
jquery.min.js:4 GET http://localhost:8080/item

It also prints out, 'an error occured in get items'.

so its doing the options request fine but something is stopping it at the get request.

Something to do with the get items function, in the parameters what you write in the body part is then appended to the end of the http. I'm just not sure that the script knows that the items it is trying to return are those items in the dictionary and I believe that's what the issue is