reindexes but doesn't shift ID's
I need to gneeralise this for any item that hte user wants to delete instead of just the item with index 1. In my head the best way to do this is to use the ID. so I am trying to get the ID of the deleted item, and subtract 1 from the ID of all teh items after this deleted item. then if the indexes all match up, I can generalise using the splice method using the ID's (not sure how you would do this with the index)/ => think this will work as line 101 in server.js returns the correct id of the element i want to delete. but teh problem is i still don't think I've getting into the delete make request (line 236 in script.js) which is making it hard to pull the ID.